Add use-service to master role
This commit adds `use-service=true` to the master role. This value
is required for the `with_puppet_running_on` beaker method to use
the deployed `puppetserver` instance as the master rather than
attempting to spin up a passenger based master.

In the `puppet` aio acceptance task, this setting is applied with
[this pre-suite](https://github.com/puppetlabs/puppet/blob/master/acceptance/setup/aio/pre-suite/045_EnsureMasterStartedOnPassenger.rb).
This commit is intended to bring the puppetstein bootstrapping
process in parity with that step.