Use tests from topic branch if provided
This commit updates the logic used to clone down
component repos for tests. Previously, puppetlabs:master
was always used, but a topic branch may include new acceptance
tests. This commit ensures that if a topic branch is provided
for a component and tests are requested from that component,
then we'll clone that fork/branch to acquire tests.